pyproject.toml
[project]
name = "todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-socketio>=5.5.1",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9.10",
]
